requirements.txt
Flask
joblib
numpy
twilio
python-dotenv